requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Jinja2==3.1.3
MarkupSafe==2.1.5
PyJWT==2.8.0
Werkzeug==3.0.2
aniso8601==9.0.1
attrs==23.2.0
blinker==1.8.1
click==8.1.7
colorama==0.4.6
flask-restx==1.3.0
importlib_resources==6.4.0
itsdangerous==2.2.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mysql-connector-python==8.4.0
simplejson==3.19.2
waitress==3.0.0
